streamable: Merge `strictdataclass` into `Streamable` class
